+++ docs/sections/changelog.rst
@@ -8,6 +8,11 @@ All notable changes to this project will be documented in this file.
 Unreleased
 ==========
 
+* Added: scope on token and introspection endpoints.
+* Changed: Use static instead of deprecated staticfiles template tag.
+* Fixed: example project on Django 2.1.
+* Fixed: example in docs for translatable scopes (ugettext).
+
 0.7.0
 =====